출력창 v1
